pyproject.toml
[tool.poetry]
name = "insightly"
version = "0.1.0"
description = ""
authors = ["Patrick Walsh <patrick.walsh@artemislab.org>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.2.4"
pandas = "^2.2.3"
plotly = "^6.0.1"
duckdb = "^1.2.2"
deltalake = "^0.25.5"
arrow = "^1.3.0"
python-dotenv = "^1.1.0"
pyspark = "^3.5.5"
setuptools = "^78.1.0"
delta-spark = "^3.3.0"
boto3 = "^1.37.37"
botocore = "^1.37.37"
sqlalchemy = "^2.0.40"
ipykernel = "^6.29.5"
langchain = "^0.3.24"
langchain-openai = "^0.3.14"
langgraph = "^0.3.34"
pre-commit = "^4.2.0"
grandalf = "^0.8"
nbformat = "^5.10.4"
fastapi = { extras = ["all"], version = "^0.115.12" }
supabase = "^2.15.2"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
